+++ docs/pyagentspec/source/changelog.rst
@@ -7,6 +7,12 @@ Agent Spec |release|
 New features
 ^^^^^^^^^^^^
 
+* **ToolBoxes**
+
+  Toolboxes are now available in the Agent Spec Language Specification and can be
+  passed to :ref:`Agents <agent>`. For more information read the :doc:`API Reference <api/tools>`.
+
+
 * **Structured Generation**
 
   Formally introduced Structured Generation in the Agent Spec Language Specification.